requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
bokeh==3.3.3
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
catalogue==2.0.10
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1702070282333/work
curl-cffi==0.5.10
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==4.1.1
effdet==0.4.1
einops==0.7.0
emoji==2.9.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
Faker==22.2.0
fastapi==0.106.0
favicon==0.7.0
feedparser==6.0.10
filelock @ file:///croot/filelock_1700591183607/work
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fschat==0.2.34
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
greenlet==3.0.3
gTTS==2.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
hyperframe==6.0.1
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.11.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.0.344
langchain-core==0.0.13
langchain-experimental==0.0.43
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.75
layoutparser==0.3.4
llama-index==0.9.29
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.4.12
markdownify==0.11.6
markdownlit==0.0.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
metaphor-python==0.1.23
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
more-itertools==10.2.0
mpmath @ file:///croot/mpmath_1690848262763/work
msg-parser==1.2.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.6.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow @ file:///croot/pillow_1696580024257/work
pluggy==1.3.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
pyarrow==14.0.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.7
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pypandoc==1.12
pyparsing==3.1.1
pypdfium2==4.25.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
rapidfuzz==3.6.0
rapidocr-onnxruntime==1.3.9
ray==2.9.0
referencing==0.32.0
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
rich==13.7.0
rpds-py==0.15.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
shapely==2.0.2
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.19
srsly==2.4.8
sse-starlette==1.8.2
st-annotated-text==4.0.1
starlette==0.27.0
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3
streamlit-antd-components==0.2.5
streamlit-authenticator==0.2.3
streamlit-bokeh-events==0.1.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-chatbox==1.1.11
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-feedback==0.1.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-modal==0.1.0
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
strsimpy==0.2.1
svgwrite==1.4.3
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
transformers-stream-generator==0.0.4
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.0
unstructured-inference==0.7.15
unstructured.pytesseract==0.3.12
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
vllm==0.2.2
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.12
weasel==0.3.4
websockets==12.0
wrapt==1.16.0
xformers==0.0.22.post7
xlrd==2.0.1
XlsxWriter==3.1.9
xyzservices==2023.10.1
yarl==1.9.4
youtube-search==2.1.2
zipp==3.17.0
